Removed innecesary code

+++ src/components/AccountsoperationsCU.vue
@@ -2,17 +2,16 @@
     <div>
         <h1>{{dialog_title_ao()}}</h1>
         <v-form ref="form_ao" v-model="form_valid_ao" lazy-validation>
-            <v-autocomplete :readonly="deleting" autoindex="3" :items="$store.state.accounts.filter(v =>v.active==true)" v-model="newao.accounts" :label="$t('Select an account')" item-text="localname" item-value="url" :rules="RulesSelection(true)" @change="on_account_change"></v-autocomplete>
-            <MyDateTimePicker :readonly="deleting" autoindex="4" label="Select operation date and time" v-model="newao.datetime" :rules="RulesDatetime(true)"></MyDateTimePicker>
-            <v-autocomplete :readonly="deleting" autoindex="0" autofocus :items="$store.state.concepts" v-model="newao.concepts" :label="$t('Select a concept')" item-text="localname" item-value="url" :rules="RulesSelection(true)"></v-autocomplete>
-            <v-text-field :readonly="deleting" autoindex="1" v-model="newao.amount"  :label="$t('Operation amount')" :placeholder="$t('Account number')" :rules="RulesFloat(30,true,this.account.decimals)" counter="30"/>
-            <v-text-field :readonly="deleting" autoindex="2" v-model="newao.comment" type="text" :label="$t('Operation comment')" :placeholder="$t('Operation comment')" counter="200"/>
+            <v-autocomplete :readonly="mode=='D'" autoindex="3" :items="$store.state.accounts.filter(v =>v.active==true)" v-model="newao.accounts" :label="$t('Select an account')" item-text="localname" item-value="url" :rules="RulesSelection(true)" @change="on_account_change"></v-autocomplete>
+            <MyDateTimePicker :readonly="mode=='D'" autoindex="4" label="Select operation date and time" v-model="newao.datetime" :rules="RulesDatetime(true)"></MyDateTimePicker>
+            <v-autocomplete :readonly="mode=='D'" autoindex="0" autofocus :items="$store.state.concepts" v-model="newao.concepts" :label="$t('Select a concept')" item-text="localname" item-value="url" :rules="RulesSelection(true)"></v-autocomplete>
+            <v-text-field :readonly="mode=='D'" autoindex="1" v-model="newao.amount"  :label="$t('Operation amount')" :placeholder="$t('Account number')" :rules="RulesFloat(30,true,this.account.decimals)" counter="30"/>
+            <v-text-field :readonly="mode=='D'" autoindex="2" v-model="newao.comment" type="text" :label="$t('Operation comment')" :placeholder="$t('Operation comment')" counter="200"/>
         </v-form>
         <v-card-actions>
             <v-spacer></v-spacer> 
-            <v-btn v-if="deleting" color="error" @click="deleteAO()" :disabled="!form_valid_ao">{{ $t("Delete") }}</v-btn>
-            <v-btn v-if="!deleting" color="primary" @click="following_ao=false;acceptDialogAO()" :disabled="!form_valid_ao">{{ button() }}</v-btn>
-            <v-btn v-if="!deleting && !editing" color="primary" @click="following_ao=true;acceptDialogAO()" :disabled="!form_valid_ao" >{{ $t("Add and follow") }}</v-btn>
+            <v-btn color="primary" @click="following_ao=false;acceptDialogAO()" :disabled="!form_valid_ao">{{ button() }}</v-btn>
+            <v-btn v-if="!mode=='D' && !mode=='U'" color="primary" @click="following_ao=true;acceptDialogAO()" :disabled="!form_valid_ao" >{{ $t("Add and follow") }}</v-btn>
         </v-card-actions>
     </div>
 </template>
@@ -28,11 +27,9 @@
             ao:{
                 required:true,
             },
-            deleting:{
-                required:false,
-                default:false
+            mode:{ // CRUD
+                required:true,
             }
-            
         },
         data () {
             return {
@@ -40,7 +37,6 @@
                 newao:null,
                 form_valid_ao:true,
                 following_ao:false,
-                editing:false,
             }
         },
         methods:{
@@ -60,14 +56,14 @@
                 }
 
                 //Accept
-                if (this.editing==true){               
+                if (this.mode=='U'){               
                     axios.put(this.newao.url, this.newao, this.myheaders())
                     .then(() => {
                             this.$emit('cruded', this.following_ao)
                     }, (error) => {
                         this.parseResponseError(error)
                     })
-                } else{
+                } else if (this.mode=='C'){ 
                     axios.post(`${this.$store.state.apiroot}/api/accountsoperations/`, this.newao,  this.myheaders())
                     .then(() => {             
                         if (this.following_ao==true){
@@ -80,37 +76,38 @@
                     }, (error) => {
                         this.parseResponseError(error)
                     })
+                } else if (this.mode=='D'){
+                    var r
+                    r = confirm(this.$t("Do you want to delete this account operation?"))
+                    if(r == false) {
+                        return
+                    }  
+                    this.following_ao=false
+                    axios.delete(this.newao.url, this.myheaders())
+                    .then(() => {
+                        this.$emit('cruded', this.following_ao)
+                    }, (error) => {
+                        this.parseResponseError(error)
+                    });
                 }
             },
             button(){
-                if(this.editing==true){
+                if (this.mode=="U"){
                     return this.$t("Update")
-                } else {
+                } else  if (this.mode=="C"){
                     return this.$t("Create")
+                } else  if (this.mode=="D"){
+                    return this.$t("Delete")
                 }
             },
             dialog_title_ao(){
-                if (this.deleting==true){
+                if (this.mode=='D'){
                     return this.$t("Deleting account operation")
-                } else if (this.editing==true){
+                } else if (this.mode=='U'){
                     return this.$t("Updating account operation")
                 } else {
                     return this.$t("Creating a new account operation")
                 }
-            },
-            deleteAO(){
-                var r
-                r = confirm(this.$t("Do you want to delete this account operation?"))
-                if(r == false) {
-                    return
-                }  
-                this.following_ao=false
-                axios.delete(this.newao.url, this.myheaders())
-                .then(() => {
-                    this.$emit('cruded', this.following_ao)
-                }, (error) => {
-                    this.parseResponseError(error)
-                });
             },
             on_account_change(){
                 this.account=this.$store.getters.getObjectByUrl("accounts",this.newao.accounts)
@@ -118,12 +115,6 @@
             }
         },
         created(){
-
-            if ( this.ao.url!=null){ // EDITING TIENE IO URL
-                this.editing=true
-            } else { // NEW IO BUT SETTING VALUES WITH URL=null
-                this.editing=false
-            }
             this.newao=Object.assign({},this.ao)
             this.on_account_change() //Updates  account object
         }
